Added seasonal_analysis.py
seasonal_analysis.py finds and predicts the popular itemsets for each season using a decision tree model with RFE feature selection and one-hot encoding.
The predicted and actual top 10 itemsets with more than 1 item for each season are then displayed in a graph, but commented out print statements can be used to display every predicted and real itemset and the number of times it was purchased for each season in descending order.